+++ app/src/main/java/com/hongstudio/parsing_github_repositories/ui/bindingadapters/ImageViewBindingAdapter.kt
@@ -1,6 +1,5 @@
-package com.hongstudio.parsing_github_repositories.util
+package com.hongstudio.parsing_github_repositories.ui.bindingadapters
 
-import android.view.View
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
@@ -9,9 +8,3 @@ import com.bumptech.glide.Glide
 fun ImageView.imageUrl(url: String?) {
     Glide.with(this).load(url).error(android.R.color.transparent).into(this)
 }
-
-@BindingAdapter("isVisible")
-fun View.isVisible(isVisible: Boolean) {
-    visibility = if (isVisible) View.VISIBLE else View.GONE
-}
-
