+++ app/src/main/java/com/hongstudio/parsing_github_repositories/ui/bindingadapters/ViewBindingAdapter.kt
@@ -1,0 +1,10 @@
+package com.hongstudio.parsing_github_repositories.ui.bindingadapters
+
+import android.view.View
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("isVisible")
+fun View.isVisible(isVisible: Boolean) {
+    visibility = if (isVisible) View.VISIBLE else View.GONE
+}
+
