+++ app/src/main/java/com/hongstudio/parsing_github_repositories/ui/views/DetailActivity.kt
@@ -1,4 +1,4 @@
-package com.hongstudio.parsing_github_repositories.view
+package com.hongstudio.parsing_github_repositories.ui.views
 
 import android.content.Context
 import android.content.Intent
@@ -10,11 +10,11 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.hongstudio.parsing_github_repositories.R
+import com.hongstudio.parsing_github_repositories.data.remote.RepoModel
 import com.hongstudio.parsing_github_repositories.databinding.ActivityDetailBinding
-import com.hongstudio.parsing_github_repositories.model.RepoModel
+import com.hongstudio.parsing_github_repositories.ui.viewmodels.DetailViewModel
 import com.hongstudio.parsing_github_repositories.util.EventObserver
-import com.hongstudio.parsing_github_repositories.util.showToast
-import com.hongstudio.parsing_github_repositories.viewmodel.DetailViewModel
+import com.hongstudio.parsing_github_repositories.util.extension.showToast
 
 class DetailActivity : AppCompatActivity() {
     private lateinit var binding: ActivityDetailBinding
